Applied Charles's Feedback
Used Charles's feedback to fix setSleepMode and sleepSimple's parameter typing. Also fixed setSleepMode to not affect the SEN bit. Finally, the enterSleep function checks all applicable Serials on a board and then flushes them.
