+++ megaavr/cores/dxcore/LowPower.h
@@ -6,22 +6,18 @@ https://github.com/jnbrauer/SER401-2022-Group39-Documentation/blob/MALS-70-low-p
 #ifndef LOW_POWER_H
 #define LOW_POWER_H
 
-#ifdef __cplusplus
-extern "C"{
-#endif
-
-#define POWER_DOWN_MODE SLPCTRL_SMODE_t::SLPCTRL_SMODE_PDOWN_gc
-#define STANDBY_MODE SLPCTRL_SMODE_t::SLPCTRL_SMODE_STDBY_gc
-#define IDLE_MODE SLPCTRL_SMODE_t::SLPCTRL_SMODE_IDLE_gc
-
 //////////////////////////////////////////////////////////////
 //                  General Sleep API                       //
 //////////////////////////////////////////////////////////////
 
 /*
 Sets the sleep mode to the specified mode. Does NOT automatically enter sleep.
+Defined Sleep Modes in megaavr/extras/ioheaders/ioavr128da48.h:
+    SLEEP_MODE_IDLE
+    SLEEP_MODE_STANDBY
+    SLEEP_MODE_PWR_DOWN
 */
-void setSleepMode(uint8_t sleepMode);
+void setSleepMode(SLPCTRL_SMODE_t sleepMode);
 
 /*
 Enables sleep mode. Does NOT automatically enter sleep.
@@ -37,13 +33,13 @@ void disableSleepMode();
 Enters sleep mode. Sleep mode MUST be enabled and the mode must be set prior to calling this function.
 Flushes Serial buffer before entering sleep if flushSerial is set to true.
 */
-void enterSleep(bool flushSerial);
+void enterSleep(bool flushSerial = true);
 
 /*
 One step sleep function. Sets the sleep mode, enables sleep, and enters sleep.
 If flushSerial is true, also flushes the Serial buffer before sleeping.
 */
-void sleepSimple(uint8_t sleepMode, bool flushSerial);
+void sleepSimple(SLPCTRL_SMODE_t sleepMode, bool flushSerial = true);
 
 
 //////////////////////////////////////////////////////////////
@@ -106,9 +102,5 @@ Disable the BOD (Brown Out Detection) for reduced power consumption. The system 
 brown-out scenario.
 */
 void bodSetDisabled();
-
-#ifdef __cplusplus
-}
-#endif
 
 #endif
